implement the selection of color in dynamic coloring program
svn path=/trunk/koffice/; revision=751692

+++ krita/plugins/paintops/dynamicbrush/programs/basicshape/kis_basic_dynamic_program.cc
@@ -93,7 +93,7 @@ KisDynamicScattering KisBasicDynamicProgram::scattering( const KisPaintInformati
     if( m_enableCout )
     {
         double v = jitter( m_countJitter, m_countSensor->parameter( info ) );
-        count = ceil( m_countCount * v );
+        count = (int)ceil( m_countCount * v );
     }
     return KisDynamicScattering(count, distance);
 }
